ScenePackage
This is a simple package from renamed scene and is for walk through on changes in environment to be reviewed and decided upon for use in game. Due to lack of understanding in GIT uses I just made package to prevent any influence on the comitt push. This content is part of group assignment in circuit stream education and is only for the explicit members of group/circuitstream.